+++ src/main/kotlin/no/nav/eessi/pensjon/personidentifisering/PersonidentifiseringService.kt
@@ -8,18 +8,10 @@ import no.nav.eessi.pensjon.json.toJson
 import no.nav.eessi.pensjon.models.BucType
 import no.nav.eessi.pensjon.models.HendelseType
 import no.nav.eessi.pensjon.models.Saktype
-import no.nav.eessi.pensjon.models.sed.kanInneholdeIdentEllerFdato
 import no.nav.eessi.pensjon.personidentifisering.helpers.FodselsdatoHelper
 import no.nav.eessi.pensjon.personidentifisering.helpers.PersonSok
 import no.nav.eessi.pensjon.personidentifisering.helpers.SedFnrSok
-import no.nav.eessi.pensjon.personidentifisering.relasjoner.GenericRelasjon
-import no.nav.eessi.pensjon.personidentifisering.relasjoner.P15000Relasjon
-import no.nav.eessi.pensjon.personidentifisering.relasjoner.P2000Relasjon
-import no.nav.eessi.pensjon.personidentifisering.relasjoner.P2100Relasjon
-import no.nav.eessi.pensjon.personidentifisering.relasjoner.P6000Relasjon
-import no.nav.eessi.pensjon.personidentifisering.relasjoner.P8000AndP10000Relasjon
-import no.nav.eessi.pensjon.personidentifisering.relasjoner.R005Relasjon
-import no.nav.eessi.pensjon.personidentifisering.relasjoner.T2000TurboRelasjon
+import no.nav.eessi.pensjon.personidentifisering.relasjoner.RelasjonsHandler
 import no.nav.eessi.pensjon.personoppslag.Fodselsnummer
 import no.nav.eessi.pensjon.personoppslag.pdl.PersonService
 import no.nav.eessi.pensjon.personoppslag.pdl.model.AdressebeskyttelseGradering
@@ -74,7 +66,7 @@ class PersonidentifiseringService(
         erNavCaseOwner: Boolean = false
     ): IdentifisertPerson? {
 
-        val potensiellePersonRelasjoner = hentRelasjoner(sedListe, bucType)
+        val potensiellePersonRelasjoner = RelasjonsHandler().hentRelasjoner(sedListe, bucType)
 
         val identifisertePersoner = hentIdentifisertePersoner(sedListe, bucType, potensiellePersonRelasjoner, hendelsesType, rinaDocumentId)
 
@@ -94,50 +86,6 @@ class PersonidentifiseringService(
         personSok.sokPersonEtterFnr(potensiellePersonRelasjoner, rinaDocumentId, bucType, sedType, hendelsesType)
         ?.let { personRelasjon -> return hentIdentifisertPerson(personRelasjon, hendelsesType) }
 
-        return null
-    }
-
-    fun hentRelasjoner(seder: List<Pair<String, SED>>, bucType: BucType): List<SEDPersonRelasjon> {
-            val fnrListe = mutableSetOf<SEDPersonRelasjon>()
-            val sedMedForsikretPrioritet = listOf(SedType.H121, SedType.H120, SedType.H070)
-
-            seder.forEach { (rinaDocumentId,sed) ->
-                try {
-                    getRelasjonHandler(sed, bucType, rinaDocumentId)?.let { handler ->
-                        fnrListe.addAll(handler.hentRelasjoner())
-                    }
-
-                } catch (ex: Exception) {
-                    logger.warn("Noe gikk galt under innlesing av fnr fra sed", ex)
-                }
-            }
-
-
-        val resultat = fnrListe
-            .filter { it.erGyldig() || it.sedType in sedMedForsikretPrioritet }
-            .filterNot { it.filterUbrukeligeElemeterAvSedPersonRelasjon() }
-            .sortedBy { it.relasjon }
-
-        return resultat.ifEmpty { fnrListe.distinctBy { it.fnr } }
-
-    }
-
-    private fun getRelasjonHandler(sed: SED, bucType: BucType, rinaDocumentId: String): T2000TurboRelasjon? {
-
-        if (sed.type.kanInneholdeIdentEllerFdato()) {
-            return when (sed.type) {
-                SedType.R005 -> R005Relasjon(sed, bucType,rinaDocumentId)
-                SedType.P2000 -> P2000Relasjon(sed, bucType,rinaDocumentId)
-                SedType.P2200 -> P2000Relasjon(sed, bucType,rinaDocumentId)
-                SedType.P2100 -> P2100Relasjon(sed, bucType,rinaDocumentId)
-                SedType.P8000, SedType.P10000 -> P8000AndP10000Relasjon(sed, bucType,rinaDocumentId)
-                SedType.P6000 -> P6000Relasjon(sed, bucType,rinaDocumentId)
-                SedType.P15000 -> P15000Relasjon(sed, bucType,rinaDocumentId)
-
-                SedType.H070, SedType.H120, SedType.H121 -> GenericRelasjon(sed, bucType, rinaDocumentId)
-                else -> GenericRelasjon(sed, bucType,rinaDocumentId)
-            }
-        }
         return null
     }
